chore: fix deprecation warnings and decompose AgentsView
- Replace deprecated rodio DeviceTrait::name() with description().name()
  across audio_output.rs (3 call sites)
- Remove dead tts_cancel field from SttPipeline in stt.rs — the worker
  thread gets its own Arc clone before storage, so the field was unused
- Extract useManagedAgentActions and usePersonaActions hooks from the
  825-line AgentsView god component, reducing it to a ~280-line
  orchestrator that wires hooks to JSX (follows the existing
  useTeamActions pattern)
